requirements.txt
fastapi>=0.104.1
uvicorn>=0.31.1
pydantic>=2.11.0
python-multipart>=0.0.9
google-generativeai>=0.3.2
yfinance>=0.2.55
pandas>=2.1.4
numpy>=1.25.0
requests>=2.32.3
aiohttp>=3.9.1
anyio>=4.5.0
